+++ src/common/resource.ts
@@ -1,1 +1,10 @@
 export const IconCube = 'https://lf3-static.bytednsdoc.com/obj/eden-cn/vhfuhpxpf/cube.png';
+
+export const defaultCubeMap = [
+    'https://lf3-static.bytednsdoc.com/obj/eden-cn/vhfuhpxpf/three/cube_01/pos-x.jpg',
+    'https://lf3-static.bytednsdoc.com/obj/eden-cn/vhfuhpxpf/three/cube_01/neg-x.jpg',
+    'https://lf3-static.bytednsdoc.com/obj/eden-cn/vhfuhpxpf/three/cube_01/pos-y.jpg',
+    'https://lf3-static.bytednsdoc.com/obj/eden-cn/vhfuhpxpf/three/cube_01/neg-y.jpg',
+    'https://lf3-static.bytednsdoc.com/obj/eden-cn/vhfuhpxpf/three/cube_01/pos-z.jpg',
+    'https://lf3-static.bytednsdoc.com/obj/eden-cn/vhfuhpxpf/three/cube_01/neg-z.jpg'
+]